Mickey, try to allow decimal numbers to calculate

+++ app/src/main/java/com/example/hydroseed/CalculatePage.java
@@ -35,7 +35,7 @@ public class CalculatePage extends AppCompatActivity {
 
         Intent getAcreage = getIntent();
 
-        double inputAcres = getAcreage.getIntExtra("Acreage", 0);
+        double inputAcres = getAcreage.getDoubleExtra("Acreage", 0);
 
         inputAcres = Math.ceil(4 * inputAcres) / 4;
